NOTICKET Fixes MacOS support + some major fixes for Linux support

+++ install.sh
@@ -2,21 +2,55 @@
 
 #set -e
 
-
 sudo apt-get install -y curl software-properties-common apt-transport-https wget
 
+# Define base directories to check
+if [ -n "$CODESPACES" ]; then
+    # GitHub Codespaces environment
+    BASE_DIR="/home/codespace"
+elif [ -n "$HOME" ]; then
+    # Local environment or other standard Unix-like systems
+    BASE_DIR="$HOME"
+else
+    # Fallback if neither $CODESPACES nor $HOME is defined
+    BASE_DIR="/home/$USER"
+fi
 
-sudo rm -rf "$HOME/.nvm"
-sudo rm -rf "$HOME/nvm"
-sudo rm -rf ~/nvm
-sudo rm -rf ~/.nvm
-# For CodeSpace
-sudo rm -rf /home/codespace/nvm/current/bin/node
-sudo rm -rf /home/"$HOME"/nvm/current/bin/node
-sudo rm -rf /home/codespace/nvm
-sudo rm -rf /home/"$HOME"/nvm
-sudo rm -rf /home/codespace/.nvm
-sudo rm -rf /home/"$HOME"/.nvm
+# Check if NVM is installed or if directories ~/.nvm or ~/nvm exist using find
+if command -v nvm &> /dev/null || find "$BASE_DIR/nvm" -mindepth 1 -maxdepth 1 -print -quit || find "$BASE_DIR/.nvm" -mindepth 1 -maxdepth 1 -print -quit; then
+    echo "Node Version Manager (NVM) found on your system."
+    echo "We are about to remove NVM from your system as it can cause conflicts with our Node.js installation."
+    echo "Please note that this will remove NVM and all its installed versions of Node.js."
+    echo "We are going to install 'n' (a Node.js version management tool) instead of NVM."
+    echo "If you do not agree with this, please type 'no' to cancel the operation."
+
+    # Ask for confirmation
+    read -p "If you agree, type 'yes' to continue: " confirmation
+
+
+    # Check user confirmation
+    if [ "$confirmation" == "yes" ]; then
+        echo "Proceeding with removal..."
+        ## REMOVING NVM AS IT CAN CAUSE A CONFLICT
+        sudo rm -rf "$HOME/.nvm"
+        sudo rm -rf "$HOME/nvm"
+        sudo rm -rf ~/nvm
+        sudo rm -rf ~/.nvm
+        # For CodeSpace
+        sudo rm -rf /home/codespace/nvm/current/bin/node
+        sudo rm -rf /home/"$HOME"/nvm/current/bin/node
+        sudo rm -rf /home/codespace/nvm
+        sudo rm -rf /home/"$HOME"/nvm
+        sudo rm -rf /home/codespace/.nvm
+        sudo rm -rf /home/"$HOME"/.nvm
+    else
+        echo "Cancellation requested. Exiting without making changes."
+        exit 1
+    fi
+else
+    echo "Node Version Manager (NVM) is not installed on your system."
+    echo "Nothing to remove."
+fi
 
 curl -fsSL https://deb.nodesource.com/setup_20.x | sudo -E bash -
 sudo apt-get install -y nodejs
